+++ server/hotrod/src/main/java/org/infinispan/server/hotrod/HotRodServer.java
@@ -36,6 +36,7 @@ import org.infinispan.commons.marshall.Externalizer;
 import org.infinispan.commons.marshall.Marshaller;
 import org.infinispan.commons.marshall.SerializeWith;
 import org.infinispan.commons.util.CollectionFactory;
+import org.infinispan.commons.util.SaslUtils;
 import org.infinispan.commons.util.ServiceFinder;
 import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.configuration.cache.Configuration;
@@ -64,7 +65,6 @@ import org.infinispan.registry.InternalCacheRegistry;
 import org.infinispan.remoting.transport.Address;
 import org.infinispan.server.core.AbstractProtocolServer;
 import org.infinispan.server.core.QueryFacade;
-import org.infinispan.server.core.security.SaslUtils;
 import org.infinispan.server.core.transport.NettyInitializers;
 import org.infinispan.server.hotrod.configuration.HotRodServerConfiguration;
 import org.infinispan.server.hotrod.counter.listener.ClientCounterManagerNotificationManager;
